refactor: replace collect_dataset_revisions with dataset_revision_collection and extend registry methods
